pyproject.toml
[tool.poetry]
name = "django-projects"
version = "0.1.1"
description = ""
authors = ["Udhay Prakash Pethakamsetty <uday3prakash@gmail.com>"]
readme = "README.md"
packages = [{include = "django_projects"}]

[tool.poetry.dependencies]
python = "^3.10.8"
django = "^4.2"
djangorestframework = "^3.14.0"
mysqlclient = "^2.1.1"
django-extensions = "^3.2.1"
pylint-django = "^2.5.3"
django-crispy-forms = "^2.0"
pandas = "^2.0.0"
requests = "^2.31.0"
pydotenv = "^0.0.7"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
pylint = "^2.17.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
